Refactor w3c color exports
