+++ mui/mui-color-constants/index.js
@@ -1,4 +1,3 @@
 export * as AlphaChannel from './alpha-channel';
-export * as CssColor from './w3c/css-named-colors';
-export * as HtmlColor from './w3c/html-named-colors';
+export { CssColor, HtmlColor } from './w3c';
 export * as MaterialDesignColor from './material-design';
