+++ mui/mui-color-constants/w3c/index.js
@@ -1,1 +1,2 @@
-export * from './css-named-colors';
+export * as CssColor from './css-named-colors';
+export * as HtmlColor from './html-named-colors';
